EventEase_server/requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
Django==5.0.6
django-appconf==1.0.6
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.1
exceptiongroup==1.2.1
Flask==3.0.3
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
MouseInfo==0.1.3
multidict==6.0.5
osmapi==4.1.0
outcome==1.3.0.post0
pillow==10.3.0
psycopg2==2.9.9
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dotenv==1.0.1
pytweening==1.2.0
pywhatkit==5.4
requests==2.31.0
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
trio==0.25.1
trio-websocket==0.11.1
twilio==9.0.5
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
wikipedia==1.4.0
wsproto==1.2.0
yarl==1.9.4
